+++ dev/inc/mpu6050.h
@@ -112,13 +112,13 @@ extern uint8 I2C_GPIO_Send_Byte(uint8 I2C_Data);
 extern uint8 I2C_GPIO_Recv_Byte(void);
 extern void I2C_GPIO_Write_Reg(uint8 I2C_Div_Adr, uint8 I2C_Reg_Adr, uint8 I2C_Data);
 extern uint8 I2C_GPIO_Read_Reg_Byte(uint8 I2C_Div_Adr, uint8 I2C_Reg_Adr);
-extern int I2C_GPIO_Read_Reg_Word(uint8 I2C_Div_Adr, uint8 I2C_Reg_Adr);
-extern int MPU_Get_Acc_X(void);
-extern int MPU_Get_Acc_Y(void);
-extern int MPU_Get_Acc_Z(void);
-extern int MPU_Get_Gyro_X(void);
-extern int MPU_Get_Gyro_Y(void);
-extern int MPU_Get_Gyro_Z(void);
+extern int16 I2C_GPIO_Read_Reg_Word(uint8 I2C_Div_Adr, uint8 I2C_Reg_Adr);
+extern int16 MPU_Get_Acc_X(void);
+extern int16 MPU_Get_Acc_Y(void);
+extern int16 MPU_Get_Acc_Z(void);
+extern int16 MPU_Get_Gyro_X(void);
+extern int16 MPU_Get_Gyro_Y(void);
+extern int16 MPU_Get_Gyro_Z(void);
 extern void MPU6050_Init(void);
 
 #endif
